requirements.txt
# 음성 파이프라인 필수 패키지
torch>=1.9.0
torchaudio>=0.9.0
speechbrain>=0.5.0
openai-whisper>=20230314
soundfile>=0.10.0
librosa>=0.8.0
numpy>=1.21.0
scipy>=1.7.0

# 추가 유틸리티
pathlib2>=2.3.0
